Написал кеш+контроллер для курса.

+++ src/main/java/dataAccess/entity/User.java
@@ -82,7 +82,7 @@ public class User implements Serializable
 		return id;
 	}
 
-	public void setId(int id)
+	public void setId(long id)
 	{
 		this.id = id;
 	}
